unify dace module
