+++ src/gt4py/next/ffront/decorator.py
@@ -306,7 +306,7 @@ class FrozenProgram:
 
 
 try:
-    from gt4py.next.program_processors.runners.dace_fieldview.program import Program
+    from gt4py.next.program_processors.runners.dace.program import Program
 except ImportError:
     pass
