+++ src/gt4py/next/program_processors/runners/dace/__init__.py
@@ -1,0 +1,27 @@
+# GT4Py - GridTools Framework
+#
+# Copyright (c) 2014-2024, ETH Zurich
+# All rights reserved.
+#
+# Please, refer to the LICENSE file in the root directory.
+# SPDX-License-Identifier: BSD-3-Clause
+
+
+from gt4py.next.program_processors.runners.dace.gtir_sdfg import build_sdfg_from_gtir
+from gt4py.next.program_processors.runners.dace.sdfg_callable import get_sdfg_args
+from gt4py.next.program_processors.runners.dace.workflow import (
+    run_dace_cpu,
+    run_dace_cpu_noopt,
+    run_dace_gpu,
+    run_dace_gpu_noopt,
+)
+
+
+__all__ = [
+    "build_sdfg_from_gtir",
+    "get_sdfg_args",
+    "run_dace_cpu",
+    "run_dace_cpu_noopt",
+    "run_dace_gpu",
+    "run_dace_gpu_noopt",
+]
