+++ src/gt4py/next/program_processors/runners/dace/gtir_builtin_translators.py
@@ -23,19 +23,19 @@ from gt4py.next.iterator.ir_utils import (
     domain_utils,
     ir_makers as im,
 )
-from gt4py.next.program_processors.runners.dace_common import utility as dace_utils
-from gt4py.next.program_processors.runners.dace_fieldview import (
+from gt4py.next.program_processors.runners.dace import (
     gtir_dataflow,
     gtir_python_codegen,
     gtir_sdfg,
+    sdfg_callable_args,
     utility as dace_gtir_utils,
 )
-from gt4py.next.program_processors.runners.dace_fieldview.gtir_scan_translator import translate_scan
+from gt4py.next.program_processors.runners.dace.gtir_scan_translator import translate_scan
 from gt4py.next.type_system import type_info as ti, type_specifications as ts
 
 
 if TYPE_CHECKING:
-    from gt4py.next.program_processors.runners.dace_fieldview import gtir_sdfg
+    from gt4py.next.program_processors.runners.dace import gtir_sdfg
 
 
 def get_domain_indices(
@@ -593,7 +593,7 @@ def translate_index(
     index_node = state.add_access(index_data)
     index_value = gtir_dataflow.ValueExpr(
         dc_node=index_node,
-        gt_dtype=dace_utils.as_itir_type(INDEX_DTYPE),
+        gt_dtype=sdfg_callable_args.as_itir_type(INDEX_DTYPE),
     )
     index_write_tasklet = sdfg_builder.add_tasklet(
         "index",
@@ -666,7 +666,7 @@ def _get_symbolic_value(
     )
     temp_name, _ = sdfg.add_scalar(
         temp_name or sdfg.temp_data_name(),
-        dace_utils.as_dace_type(scalar_type),
+        sdfg_callable_args.as_dace_type(scalar_type),
         find_new_name=True,
         transient=True,
     )
@@ -805,7 +805,7 @@ def translate_scalar_expr(
             dace.Memlet(data=arg_node.data, subset="0"),
         )
     # finally, create temporary for the result value
-    temp_name, _ = sdfg_builder.add_temp_scalar(sdfg, dace_utils.as_dace_type(node.type))
+    temp_name, _ = sdfg_builder.add_temp_scalar(sdfg, sdfg_callable_args.as_dace_type(node.type))
     temp_node = state.add_access(temp_name)
     state.add_edge(
         tasklet_node,
