+++ src/gt4py/next/program_processors/runners/dace/program.py
@@ -20,7 +20,7 @@ from gt4py.next.ffront import decorator
 from gt4py.next.iterator import ir as itir, transforms as itir_transforms
 from gt4py.next.iterator.transforms import extractors as extractors
 from gt4py.next.otf import arguments, recipes, toolchain
-from gt4py.next.program_processors.runners.dace_common import utility as dace_utils
+from gt4py.next.program_processors.runners.dace import sdfg_callable_args
 from gt4py.next.type_system import type_specifications as ts
 
 
@@ -152,15 +152,15 @@ class Program(decorator.Program, dace.frontend.python.common.SDFGConvertible):
             in_arrays_with_id = [
                 (name, conn_id)
                 for name in with_table
-                if (conn_id := dace_utils.connectivity_identifier(name))
+                if (conn_id := sdfg_callable_args.connectivity_identifier(name))
                 in self.sdfg_closure_cache["arrays"]
             ]
             in_arrays = (name for name, _ in in_arrays_with_id)
             name_axis = list(itertools.product(in_arrays, [0, 1]))
 
             def size_symbol_name(name: str, axis: int) -> str:
-                return dace_utils.field_size_symbol_name(
-                    dace_utils.connectivity_identifier(name), axis
+                return sdfg_callable_args.field_size_symbol_name(
+                    sdfg_callable_args.connectivity_identifier(name), axis
                 )
 
             connectivity_tables_size_symbols = {
@@ -169,8 +169,8 @@ class Program(decorator.Program, dace.frontend.python.common.SDFGConvertible):
             }
 
             def stride_symbol_name(name: str, axis: int) -> str:
-                return dace_utils.field_stride_symbol_name(
-                    dace_utils.connectivity_identifier(name), axis
+                return sdfg_callable_args.field_stride_symbol_name(
+                    sdfg_callable_args.connectivity_identifier(name), axis
                 )
 
             connectivity_table_stride_symbols = {
@@ -196,12 +196,12 @@ class Program(decorator.Program, dace.frontend.python.common.SDFGConvertible):
                     self.connectivity_tables_data_descriptors[conn_id] = dace.data.Array(
                         dtype=dace.dtypes.dtype_to_typeclass(conn.dtype.dtype.type),
                         shape=[
-                            symbols[dace_utils.field_size_symbol_name(conn_id, 0)],
-                            symbols[dace_utils.field_size_symbol_name(conn_id, 1)],
+                            symbols[sdfg_callable_args.field_size_symbol_name(conn_id, 0)],
+                            symbols[sdfg_callable_args.field_size_symbol_name(conn_id, 1)],
                         ],
                         strides=[
-                            symbols[dace_utils.field_stride_symbol_name(conn_id, 0)],
-                            symbols[dace_utils.field_stride_symbol_name(conn_id, 1)],
+                            symbols[sdfg_callable_args.field_stride_symbol_name(conn_id, 0)],
+                            symbols[sdfg_callable_args.field_stride_symbol_name(conn_id, 1)],
                         ],
                         storage=Program.connectivity_tables_data_descriptors["storage"],
                     )
@@ -221,7 +221,7 @@ def _crosscheck_dace_parsing(dace_parsed_args: list[Any], gt4py_program_args: li
     ):
         match dace_parsed_arg:
             case dace.data.Scalar():
-                assert dace_parsed_arg.dtype == dace_utils.as_dace_type(gt4py_program_arg)
+                assert dace_parsed_arg.dtype == sdfg_callable_args.as_dace_type(gt4py_program_arg)
             case bool() | np.bool_():
                 assert isinstance(gt4py_program_arg, ts.ScalarType)
                 assert gt4py_program_arg.kind == ts.ScalarKind.BOOL
@@ -238,7 +238,9 @@ def _crosscheck_dace_parsing(dace_parsed_args: list[Any], gt4py_program_args: li
                 assert isinstance(gt4py_program_arg, ts.FieldType)
                 assert isinstance(gt4py_program_arg.dtype, ts.ScalarType)
                 assert len(dace_parsed_arg.shape) == len(gt4py_program_arg.dims)
-                assert dace_parsed_arg.dtype == dace_utils.as_dace_type(gt4py_program_arg.dtype)
+                assert dace_parsed_arg.dtype == sdfg_callable_args.as_dace_type(
+                    gt4py_program_arg.dtype
+                )
             case dace.data.Structure() | dict() | collections.OrderedDict():
                 # offset provider
                 pass
