+++ src/gt4py/next/program_processors/runners/dace/transformations/gpu_utils.py
@@ -22,9 +22,7 @@ from dace import (
 from dace.codegen.targets import cpp as dace_cpp
 from dace.sdfg import nodes as dace_nodes
 
-from gt4py.next.program_processors.runners.dace_fieldview import (
-    transformations as gtx_transformations,
-)
+from gt4py.next.program_processors.runners.dace import transformations as gtx_transformations
 
 
 def gt_gpu_transformation(
