+++ src/gt4py/next/program_processors/runners/dace/transformations/map_promoter.py
@@ -17,9 +17,7 @@ from dace import (
 )
 from dace.sdfg import nodes as dace_nodes
 
-from gt4py.next.program_processors.runners.dace_fieldview import (
-    transformations as gtx_transformations,
-)
+from gt4py.next.program_processors.runners.dace import transformations as gtx_transformations
 
 
 @dace_properties.make_properties
